requirements.txt
pytorch_lightning==1.2
transformers
Pillow
tqdm==4.56.0
ipdb==0.13.4
numpy
einops==0.3.0
#pyarrow==2.0.0
sacred
pandas
git+https://github.com/rwightman/pytorch-image-models.git